Created PurchaseService and its implementation, added total to Invoice, currency values converted to BigDecimal

+++ src/main/java/com/ucx/domainexercise/domain/Product.java
@@ -5,15 +5,17 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class Product extends BaseModel<Integer> {
     private String name;
-    private Double unitPrice;
+    private BigDecimal unitPrice;
     private Boolean inStock;
 
     @Builder
-    public Product(Integer id, RecordStatus recordStatus, String name, Double unitPrice, Boolean inStock) {
+    public Product(Integer id, RecordStatus recordStatus, String name, BigDecimal unitPrice, Boolean inStock) {
         super(id, recordStatus);
         this.name = name;
         this.unitPrice = unitPrice;
